Use newer docker cli commands
This change is a bit opinionated because the "old" commands still work
and some old commands like exec, run and start are still not used in
context of container.

But for the changed commands I find it much easier to remember them when
using the image, container, volume, network sub commands. Especially
because their syntax is now very similar. For example to list images,
container, volumes and networks you just type

    docker image ls
    docker container ls
    docker volume ls
    docker network ls

instead of
    docker images
    docker ps